refactor: getPosition() 제거 및 로직 수정 getPosition() 메소드 로직으로 대체
